add type-safe button groups
The four kinds of buttons (normal, link, radio, checkbox) all have their
own opaque wrapper type. This ensures that only one kind of button
exists per button group. Composition of different button kinds into one
toolbar is still possible.

An unadded feature is the `name` attribute for a group of radio buttons.

Radio buttons are represented as a label enclosing a radio input

    <div role="group" data-toggle="buttons" class="btn-group">
        <label class="active btn btn-secondary" data-toggle="button">
            <input type="radio" autocomplete="off">On
        </label>
        <label class=" btn btn-secondary" data-toggle="button">
            <input type="radio" autocomplete="off">Off
        </label>
    </div>

The problem is that the bootstrap attributes must be applied to the
surrounding label. Thus, the Options that `Button.radioButton` gets as
an argument are for the label, not for the input. Therefore, it is
impossible to set the name attribute for the `input` element.

I'm not sure whether this is a problem for elm. In practice, a radio
button group will be driven by some sum type, for instance

    type Light = On | Off

your model will contain this value

    type alias Model = { light : Light }

So, regardless of the dom, it is impossible to have more than 2 or fewer
than 1 radio button selected, because such a state cannot be expressed.

So I suggest we don't implement this for now, and see whether we can
find a use case for "one name for all radios in a group" within an elm
application.