Support having only one card open at any time for accordion
One caveat is that if you use withAnimation and start with one card open the animation looks pants.
Reason being that we don't have the height for the open card when you click another card (or self).
With 0.19 this should be possible to amend because we can then get element dimensions through commands/tasks and not only through dom events.

This fixes #66 and fixes #115.