Allows an action on tab select
There was no way to do an action on tab selection, only on tab click.
This was problematic because tabs are selected on pageload, which means
the first tab would be expanded without calling the action if you rely
on tab click. Our specific usecase is to make an API call on tab select
rather than on tab click, and this commit makes that possible.

In order to not break the existing API we created a new function,
`viewWithAction`, instead of modifying `view`.